fix: Prevent deletion of the admin team #2589
Approved-by: Titouan-Joseph CICORELLA <titouanjoseph@gmail.com>

Closes #2589

See merge request 24-heures-insa/overbookd-mono!3621

+++ apps/api/src/team/team.service.ts
@@ -67,6 +67,9 @@ export class TeamService {
   }
 
   async deleteTeam(code: string): Promise<void> {
+    if (code === ADMIN) {
+      throw new UnauthorizedException("Touche pas à l'équipe admin !");
+    }
     await this.prisma.team.delete({ where: { code } });
   }
